requirements.txt
requests
python-dotenv
confluent-kafka
pymongo
pymongo